Added new folder for my comprehension of how the program works, with the program.py being a copy of asignmenta_2.py and then functions.py being a edited copy of modified_test.py with error preventions built in to each function via while loops and if statement. this is changed from the prior faulty way of rerunning a function pointed out earlier by a comment.
I also added a removable easter egg to the edit function which hides options to edit the id, gpa, and semester. Also for testing purposes I added a few starter students which come before the id 123 for keeping it in line with his test shown.

I attempted to make the modified_test.py also line up with the 'don't rerun functions as error prevention' guideline.

if any changes are unpleasent or wrong let me know so I can revert or fix them.